+++ G9SuperNetCoreServer/G9Common/Abstract/ASessionHandler.cs
@@ -32,21 +32,21 @@ namespace G9Common.Abstract
 
         /// <summary>
         ///     Access to action send command by name
-        ///     long => session id
+        ///     uint => session id
         ///     string => command name
         ///     object => data for send
         ///     Return => int number specify byte to send. if don't send return 0
         /// </summary>
-        public Func<long, string, object, int> SendCommandByName;
+        public Func<uint, string, object, int> SendCommandByName;
 
         /// <summary>
         ///     Access to action send command by name async
-        ///     long => session id
+        ///     uint => session id
         ///     string => command name
         ///     object => data for send
         ///     Return => Task int number specify byte to send. if don't send return 0
         /// </summary>
-        public Func<long, string, object, Task<int>> SendCommandByNameAsync;
+        public Func<uint, string, object, Task<int>> SendCommandByNameAsync;
         #endregion
 
         #endregion
