+++ G9SuperNetCoreServer/G9SuperNetCoreClient/Helper/G9SuperNetCoreStateObjectClient.cs
@@ -5,7 +5,7 @@ namespace G9SuperNetCoreClient.Helper
     // State object for reading client data asynchronously  
     internal class G9SuperNetCoreStateObjectClient : AG9SuperNetCoreStateObjectBase
     {
-        public G9SuperNetCoreStateObjectClient(int oBufferSize, long oSessionIdentity) : base(oBufferSize, oSessionIdentity)
+        public G9SuperNetCoreStateObjectClient(int oBufferSize, uint oSessionIdentity) : base(oBufferSize, oSessionIdentity)
         {
         }
     }
