+++ G9SuperNetCoreServer/G9SuperNetCoreServer/HelperClass/G9SuperNetCoreStateObjectServer.cs
@@ -5,7 +5,7 @@ namespace G9SuperNetCoreServer.HelperClass
     // State object for reading client data asynchronously  
     internal class G9SuperNetCoreStateObjectServer : AG9SuperNetCoreStateObjectBase
     {
-        public G9SuperNetCoreStateObjectServer(int oBufferSize, long oSessionIdentity) : base(oBufferSize, oSessionIdentity)
+        public G9SuperNetCoreStateObjectServer(int oBufferSize, uint oSessionIdentity) : base(oBufferSize, oSessionIdentity)
         {
         }
     }
